增加自动榜单“最新更新王国”，并ims系统可设置

+++ content-api/src/main/java/com/me2me/content/dto/BillBoardDetailsDto.java
@@ -40,7 +40,7 @@ public class BillBoardDetailsDto implements BaseEntity {
     @Data
     public static class InnerDetailData implements BaseEntity{
 
-    	private int sinceId;
+    	private long sinceId;
     	
         private int subType;
